refactor: 영화 여러 개의 정보 fetcher를 별도의 파일로 분리
